move util functions out of main logic file
