+++ mpldxf/backend_dxf.py
@@ -55,7 +55,8 @@ import ezdxf
 from ezdxf.enums import TextEntityAlignment
 from ezdxf.math.clipping import ClippingRect2d
 
-from . import dxf_colors
+from .color_utils import rgb_to_dxf
+from .geometry_utils import filter_invalid_coordinates, is_valid_coordinate
 
 # When packaged with py2exe ezdxf has issues finding its templates
 # We tell it where to find them using this.
@@ -63,37 +64,6 @@ from . import dxf_colors
 # configuration in setup.py
 if hasattr(sys, "frozen"):
     ezdxf.options.template_dir = os.path.dirname(sys.executable)
-
-
-def rgb_to_dxf(rgb_val):
-    """Convert an RGB[A] colour to DXF colour index."""
-    if rgb_val is None:
-        dxfcolor = dxf_colors.WHITE
-    # change black to white
-    elif np.allclose(np.array(rgb_val[:3]), np.zeros(3)):
-        dxfcolor = dxf_colors.nearest_index([255, 255, 255])
-    else:
-        dxfcolor = dxf_colors.nearest_index([255.0 * val for val in rgb_val[:3]])
-    return dxfcolor
-
-
-def is_valid_coordinate(coord):
-    """Check if a coordinate contains only finite numbers (no NaN or Inf)."""
-    coord_array = np.asarray(coord)
-    return np.all(np.isfinite(coord_array))
-
-
-def filter_invalid_coordinates(vertices):
-    """Filter out vertices with NaN or Inf values."""
-    if len(vertices) == 0:
-        return vertices
-
-    # Handle both 1D list of coordinates and 2D array of vertices
-    vertices_array = np.asarray(vertices)
-    if vertices_array.ndim == 1:
-        return vertices if is_valid_coordinate(vertices) else []
-    else:
-        return [v for v in vertices if is_valid_coordinate(v)]
 
 
 class RendererDxf(RendererBase):
